Split core.py into multiple modules.
