+++ muckrake/services/schema_registry.py
@@ -12,7 +12,11 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
+from ducktape.services.service import Service
+
 import json
+import re
+import time
 import urllib2
 
 SCHEMA_REGISTRY_V1_JSON = "application/vnd.schemaregistry.v1+json"
@@ -36,6 +40,93 @@ GENERIC_REQUEST = "application/octet-stream"
 
 # Minimum header data necessary for using the Schema Registry REST api.
 SCHEMA_REGISTRY_DEFAULT_REQUEST_PROPERTIES = {"Content-Type": SCHEMA_REGISTRY_V1_JSON_WEIGHTED, "Accept": "*/*"}
+
+
+class SchemaRegistryService(Service):
+    def __init__(self, service_context, zk, kafka):
+        """
+        :type service_context ducktape.services.service.ServiceContext
+        :type zk: ZookeeperService
+        :type kafka: muckrake.services.kafka.KafkaService
+        """
+        super(SchemaRegistryService, self).__init__(service_context)
+        self.zk = zk
+        self.kafka = kafka
+        self.port = 8081
+
+    def start_node(self, node):
+        self.logger.info("Starting Schema Registry node %d on %s", self.idx(node), node.account.hostname)
+        template = open('templates/schema-registry.properties').read()
+        template_params = {
+            'kafkastore_topic': '_schemas',
+            'kafkastore_url': self.zk.connect_setting(),
+            'rest_port': self.port
+        }
+
+        config = template % template_params
+        if config is None:
+            template = open('templates/schema-registry.properties').read()
+            template_params = {
+                'kafkastore_topic': '_schemas',
+                'kafkastore_url': self.zk.connect_setting(),
+                'rest_port': self.port
+            }
+            config = template % template_params
+
+        node.account.create_file("/mnt/schema-registry.properties", config)
+        cmd = "/opt/schema-registry/bin/schema-registry-start /mnt/schema-registry.properties " \
+            + "1>> /mnt/schema-registry.log 2>> /mnt/schema-registry.log &"
+
+        self.logger.debug("Attempting to start node with command: " + cmd)
+        node.account.ssh(cmd)
+
+    def wait_until_alive(self, node):
+        # Wait for the server to become live
+        node.account.wait_for_http_service(self.port, headers=SCHEMA_REGISTRY_DEFAULT_REQUEST_PROPERTIES)
+
+    def stop_node(self, node, clean_shutdown=True, allow_fail=True):
+        self.logger.info("Stopping %s node %d on %s" % (type(self).__name__, self.idx(node), node.account.hostname))
+        node.account.kill_process("schema-registry", clean_shutdown, allow_fail)
+
+    def clean_node(self, node):
+        self.logger.info("Cleaning %s node %d on %s" % (type(self).__name__, self.idx(node), node.account.hostname))
+        node.account.ssh("rm -rf /mnt/schema-registry.properties /mnt/schema-registry.log")
+
+    def restart_node(self, node, wait_sec=0, clean_shutdown=True):
+        self.stop_node(node, clean_shutdown, allow_fail=True)
+        time.sleep(wait_sec)
+        self.start_node(node)
+
+    def get_master_node(self):
+        node = self.nodes[0]
+
+        cmd = "/opt/kafka/bin/kafka-run-class.sh kafka.tools.ZooKeeperMainWrapper -server %s get /schema_registry/schema_registry_master" \
+              % self.zk.connect_setting()
+
+        host = None
+        port_str = None
+        self.logger.debug("Querying zookeeper to find current schema registry master: \n%s" % cmd)
+        for line in node.account.ssh_capture(cmd):
+            match = re.match("^{\"host\":\"(.*)\",\"port\":(\d+),", line)
+            if match is not None:
+                groups = match.groups()
+                host = groups[0]
+                port_str = groups[1]
+                break
+
+        if host is None:
+            raise Exception("Could not find schema registry master.")
+
+        base_url = "%s:%s" % (host, port_str)
+        self.logger.debug("schema registry master is %s" % base_url)
+
+        # Return the node with this base_url
+        for idx, node in enumerate(self.nodes, 1):
+            if self.url(idx).find(base_url) >= 0:
+                return self.get_node(idx)
+
+    def url(self, idx=1):
+        return "http://" + self.get_node(idx).account.hostname + ":" + str(self.port)
 
 
 class RequestData(object):
