+++ muckrake/tests/sr_master_failover_test.py
@@ -13,7 +13,6 @@
 # limitations under the License.
 
 from test import SchemaRegistryFailoverTest
-from muckrake.services.schema_registry_utils import SCHEMA_REGISTRY_DEFAULT_REQUEST_PROPERTIES
 import time
 
 # Specify retry frequency and retry window
